+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+
+- Automatic creation/update of the Vector Index
+
 ## 0.2.0 - 2024-08-21
 
 ### Updated